Refactor imports in test files and utility modules for consistency
