+++ vscode/microsoft-kiota/src/commands/generate/generateClientCommand.ts
@@ -5,12 +5,12 @@ import * as vscode from "vscode";
 import { API_MANIFEST_FILE, extensionId, treeViewFocusCommand, treeViewId } from "../../constants";
 import { setGenerationConfiguration } from "../../handlers/configurationHandler";
 import { clearDeepLinkParams, getDeepLinkParams } from "../../handlers/deepLinkParamsHandler";
-import { ConsumerOperation, generateClient, generatePlugin, generationLanguageToString, getLogEntriesForLevel, KiotaLogEntry, LogLevel } from "../../kiotaInterop";
-import { getLanguageInformationForDescription } from "../../kiotaInterop/languageInformation";
+import { ConsumerOperation, generateClient, generatePlugin, generationLanguageToString, getLanguageInformationForDescription, getLogEntriesForLevel, KiotaLogEntry, LogLevel } from "../../kiotaInterop";
+import { KiotaGenerationLanguage, KiotaPluginType } from "../../kiotaInterop/enums";
 import { GenerateState, generateSteps } from "../../modules/steps/generateSteps";
 import { DependenciesViewProvider } from "../../providers/dependenciesViewProvider";
 import { OpenApiTreeProvider } from "../../providers/openApiTreeProvider";
-import { GenerationType, KiotaGenerationLanguage, KiotaPluginType } from "../../types/enums";
+import { GenerationType } from "../../types/enums";
 import { ExtensionSettings, getExtensionSettings } from "../../types/extensionSettings";
 import { GeneratedOutputState } from "../../types/GeneratedOutputState";
 import { WorkspaceGenerationContext } from "../../types/WorkspaceGenerationContext";
