+++ vscode/microsoft-kiota/src/commands/generate/generation-util.ts
@@ -2,10 +2,9 @@ import * as fs from 'fs';
 import * as vscode from "vscode";
 
 import { KIOTA_WORKSPACE_FILE, treeViewId } from "../../constants";
-import { LanguagesInformation } from "../../kiotaInterop";
+import { getLanguageInformationInternal, LanguagesInformation } from "../../kiotaInterop";
 import { OpenApiTreeProvider } from "../../providers/openApiTreeProvider";
 import { getWorkspaceJsonPath, updateTreeViewIcons } from "../../util";
-import { getLanguageInformationInternal } from "../../kiotaInterop/languageInformation";
 
 let _languageInformation: LanguagesInformation | undefined; // doesn't change over the lifecycle of the extension
