+++ vscode/microsoft-kiota/src/kiotaInterop/index.ts
@@ -1,21 +1,28 @@
 
 import { KiotaGenerationLanguage, KiotaPluginType } from './enums';
-import { generateClient } from './generateClient';
-import { generatePlugin } from './generatePlugin';
-import { getKiotaVersion } from './getKiotaVersion';
-import { migrateFromLockFile } from './migrateFromLockFile';
-import { removeClient, removePlugin } from './removeItem';
-import { searchDescription } from './searchDescription';
-import { updateClients } from './updateClients';
+import { generateClient } from './lib/generateClient';
+import { generatePlugin } from './lib/generatePlugin';
+import { getKiotaVersion } from './lib/getKiotaVersion';
+import { getManifestDetails } from './lib/getManifestDetails';
+import { getLanguageInformationForDescription, getLanguageInformationInternal } from './lib/languageInformation';
+import { migrateFromLockFile } from './lib/migrateFromLockFile';
+import { removeClient, removePlugin } from './lib/removeItem';
+import { searchDescription } from './lib/searchDescription';
+import { showKiotaResult } from './lib/showKiotaResult';
+import { updateClients } from './lib/updateClients';
 
 export {
     generateClient,
     generatePlugin,
     getKiotaVersion,
+    getLanguageInformationForDescription,
+    getLanguageInformationInternal,
+    getManifestDetails,
     migrateFromLockFile,
     removeClient,
     removePlugin,
     searchDescription,
+    showKiotaResult,
     updateClients
 };
