+++ vscode/microsoft-kiota/src/kiotaInterop/lib/generatePlugin.ts
@@ -1,8 +1,8 @@
 import * as rpc from "vscode-jsonrpc/node";
 
-import { ConsumerOperation, GenerationConfiguration, KiotaLogEntry, PluginAuthType } from ".";
-import connectToKiota from "./connect";
-import { KiotaPluginType } from "./enums";
+import { ConsumerOperation, GenerationConfiguration, KiotaLogEntry, PluginAuthType } from "..";
+import connectToKiota from "../connect";
+import { KiotaPluginType } from "../enums";
 
 interface PluginGenerationOptions {
   openAPIFilePath: string;
