+++ vscode/microsoft-kiota/src/kiotaInterop/lib/migrateFromLockFile.ts
@@ -1,7 +1,7 @@
 import * as rpc from "vscode-jsonrpc/node";
 
-import { KiotaLogEntry } from ".";
-import connectToKiota from "./connect";
+import { KiotaLogEntry } from "..";
+import connectToKiota from "../connect";
 
 export function migrateFromLockFile(lockFileDirectory: string): Promise<KiotaLogEntry[] | undefined> {
   return connectToKiota(async (connection) => {
