+++ vscode/microsoft-kiota/src/kiotaInterop/lib/updateClients.ts
@@ -1,7 +1,7 @@
 import * as rpc from "vscode-jsonrpc/node";
 
-import { KiotaLogEntry } from ".";
-import connectToKiota from "./connect";
+import { KiotaLogEntry } from "..";
+import connectToKiota from "../connect";
 
 interface UpdateClientsConfiguration {
   cleanOutput: boolean;
